requirements.txt
pandas==0.22.0
numpy==1.14.2
google_images_download==2.3.0
scipy==1.0.0
opencv_python==3.4.1.15
matplotlib==2.2.2
comtypes==1.1.4
progressbar2==3.38.0
